Introducing DhtHandle and DhtKey
These data types were defined as part of this commit:

* DhtHandle
* DhtHandle'
* DhtKey (newtype)

DhtHandle is just existential wrapper on top of DhtHandle'. It is
intended to be used by code built on top of this API, where as
DhtHandle' should be used only by concrete DHT implementations.

As a consequence of DhtKey definition we need bytestring dependency.