DhtResult now contains MVar
DHT operations will be asynchronous in general, for that purpose it is
necessary to use synchronization primitives. Another thing to consider
is the fact that exception propagation might not be working the way one
wants. In example if there is a loop that processes/sends network
messages, then exception raised in that context wouldn't be caught in
those places that are actually affected by it. For this reason we now
have following definitions:

    type DhtResult m a = m (DhtResultVar a)

    newtype DhtResultVar a =
        DhtResultVar (MVar (Either SomeException a))

Notice that SomeException data type, that is top-level wrapper for
exceptions in Haskell, is part of the result. This way it is possible to
notify caller of DHT operation about relevant exceptions raised while
processing call.

This commit also includes a lot of helper functions and smart
constructors. See source code and/or API documentation for details about
them.